:sparkles: 스텟 추가 및 시간에 따른 스텟 변경 구현
변경사항:
1. 스텟 3가지 (영양, 수분, 정신력) 데이터 및 로직 구현
2. 시간의 흐름에 따른 스텟 값 변경 구현

+++ Assets/Dev/Script/Player/PlayerControllerData.cs
@@ -17,13 +17,13 @@ namespace IndieLINY
         [SerializeField] private float _walkSpeed;
 
         [Header("지구력이 0이 되는데 걸리는 시간 (sec)")]
-        [SerializeField] private float _decreaseEndurancePerSec;
+        [SerializeField] private int _decreaseEndurancePerSec;
         [Header("지구력이 최대가 되는데 걸리는 시간 (sec)")]
-        [SerializeField] private float _increaseEndurancePerSec;
+        [SerializeField] private int _increaseEndurancePerSec;
 
-        public float DecreaseEndurancePerSec => _decreaseEndurancePerSec;
+        public int DecreaseEndurancePerSec => _decreaseEndurancePerSec;
 
-        public float IncreaseEndurancePerSec => _increaseEndurancePerSec;
+        public int IncreaseEndurancePerSec => _increaseEndurancePerSec;
 
         public float WorldInteractionRadius => _worldInteractionRadius;
